+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/controller/PlantingBatchController.java
@@ -37,7 +37,7 @@ public class PlantingBatchController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ApiResponse<PlantingBatchDto>> getById(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<PlantingBatchDto>> getById(@PathVariable Integer id) {
         return plantingBatchService.findById(id)
                 .map(data -> ResponseEntity.ok(ApiResponse.success(data)))
                 .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND)
@@ -56,7 +56,7 @@ public class PlantingBatchController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ApiResponse<PlantingBatchDto>> update(@PathVariable Long id,
+    public ResponseEntity<ApiResponse<PlantingBatchDto>> update(@PathVariable Integer id,
             @RequestBody PlantingBatchDto dto) {
         try {
             return plantingBatchService.update(id, dto)
@@ -70,7 +70,7 @@ public class PlantingBatchController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse<Void>> delete(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<Void>> delete(@PathVariable Integer id) {
         try {
             if (plantingBatchService.deleteById(id)) {
                 return ResponseEntity.ok(ApiResponse.success(null, "PlantingBatch deleted successfully"));
@@ -84,7 +84,7 @@ public class PlantingBatchController {
     }
 
     @DeleteMapping("/bulk-delete")
-    public ResponseEntity<ApiResponse<Void>> deleteMultiple(@RequestBody List<Long> ids) {
+    public ResponseEntity<ApiResponse<Void>> deleteMultiple(@RequestBody List<Integer> ids) {
         try {
             if (ids == null || ids.isEmpty()) {
                 return ResponseEntity.badRequest()
