+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/dto/InventoryItemDto.java
@@ -7,7 +7,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class InventoryItemDto {
-    private Long itemId;
+    private Integer itemId;
     private String itemName;
     private String category;
     private String unit;
